fix: rename round card details to RoundOverviewCard
